+++ packages/polythene-css-shadow/index.d.ts
@@ -27,3 +27,5 @@ export const sharedVarFns: StyleFns;
 export function addStyle(customSelector: string, customVars: StyleObject, scoping?: Scoping): void;
 
 export function getStyle(customSelector: string, customVars: StyleObject, scoping?: Scoping): Array<StyleObject>;
+
+export function addGeneralStyleToHead(): void;