+++ packages/polythene-css-tabs/index.d.ts
@@ -51,3 +51,5 @@ export const tabsLayout: StyleFn;
 export function addStyle(customSelector: string, customVars: StyleObject, scoping?: Scoping): void;
 
 export function getStyle(customSelector: string, customVars: StyleObject, scoping?: Scoping): Array<StyleObject>;
+
+export function addGeneralStyleToHead(): void;