+++ scripts/bundling/rollup.base.js
@@ -56,6 +56,7 @@ export const createConfig = () => {
           },
         ]
       }),
+      
       resolve(),
 
       commonjs({